basic/ReAct/requirements.txt
openai==0.28.0
python-dotenv>=0.19.0 